backend/requirements.txt
annotated-types
anyio
certifi
charset-normalizer
click
fastapi
h11
idna
pydantic
pydantic_core
python-multipart
requests
sniffio
starlette
typing_extensions
urllib3
uvicorn
python-dotenv
numpy
pandas
joblib
scikit-learn
torch
xgboost
pillow